repeatable actions!
actions can be set as repeatable by adding a bool field in the quest called "Repeatable" and setting it to true.

if there is a field called "Points Repeat", then every time the quest runs, it will multiply the amount of points available by that field.

check the "Hotel/Action/Breakfast" quest to see it working